trees: Switch to int16_t for features [BREAKING]
This enables fixed-point arithmetics for "loadable" inference strategy,
and reduces the size of each node from 12->8 bytes on 32-bit aligned platforms

+++ emlearn/eml_trees.h
@@ -16,7 +16,7 @@ extern "C" {
 
 typedef struct _EmlTreesNode {
     int8_t feature;
-    float value;
+    int16_t value;
     int16_t left;
     int16_t right;
 } EmlTreesNode;
@@ -68,15 +68,15 @@ Returns an offset into the leaves structure
 */
 static int32_t
 eml_trees_predict_tree(const EmlTrees *forest, int32_t tree_root,
-                    const float *features, int8_t features_length)
+                        const int16_t *features, int8_t features_length)
 {
     int32_t node_idx = tree_root;
 
     // TODO: see if using a pointer node instead of indirect adressing using node_idx improves perf
     while (node_idx >= 0) {
         const int8_t feature = forest->nodes[node_idx].feature;
-        const float value = features[feature];
-        const float point = forest->nodes[node_idx].value;
+        const int16_t value = features[feature];
+        const int16_t point = forest->nodes[node_idx].value;
         //printf("node %d feature %d. %d < %d\n", node_idx, feature, value, point);
         const int16_t child = (value < point) ? forest->nodes[node_idx].left : forest->nodes[node_idx].right;
         if (child >= 0) {
@@ -100,7 +100,7 @@ eml_trees_outputs_proba(const EmlTrees *self)
 
 EmlError
 eml_trees_predict_proba(const EmlTrees *self,
-            const float *features, int8_t features_length,
+            const int16_t *features, int8_t features_length,
             float *out, int32_t out_length)
 {
     EML_PRECONDITION(features, EmlUninitialized);
@@ -174,7 +174,7 @@ eml_trees_predict_proba(const EmlTrees *self,
 * \return The class number, or -EmlTreesError on failure
 */
 int32_t
-eml_trees_predict(const EmlTrees *forest, const float *features, int8_t features_length)
+eml_trees_predict(const EmlTrees *forest, const int16_t *features, int8_t features_length)
 {
     EML_LOG_BEGIN("eml-trees-predict-start");
     EML_LOG_ADD_INTEGER("classes", forest->n_classes);
@@ -233,7 +233,7 @@ eml_trees_predict(const EmlTrees *forest, const float *features, int8_t features
 */
 EmlError
 eml_trees_regress(const EmlTrees *forest,
-        const float *features, int8_t features_length,
+        const int16_t *features, int8_t features_length,
         float *out, int8_t out_length)
 {
 
@@ -283,7 +283,7 @@ eml_trees_regress(const EmlTrees *forest,
 */
 float
 eml_trees_regress1(const EmlTrees *forest,
-        const float *features, int8_t features_length)
+        const int16_t *features, int8_t features_length)
 {
     float out[1];
     EmlError err = eml_trees_regress(forest,
